Refactoring: MapCollection
