+++ app/frontend/ui/Maps.jsx
@@ -1,14 +1,11 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { CircularProgress } from 'material-ui/Progress';
 import MapIcon from 'material-ui-icons/Map';
 import Button from 'material-ui/Button';
 import AddIcon from 'material-ui-icons/Add';
-import LockIcon from 'material-ui-icons/Lock';
-import GroupIcon from 'material-ui-icons/Group';
-import PersonIcon from 'material-ui-icons/Person';
-import GridList, { GridListTile, GridListTileBar } from 'material-ui/GridList';
 import Typography from 'material-ui/Typography';
 import SwipeableViews from 'react-swipeable-views';
+import MapCollectionContainer from '../containers/MapCollectionContainer';
 
 const styles = {
   rootLarge: {
@@ -19,16 +16,7 @@ const styles = {
     margin: '120px auto 64px'
   },
   container: {
-    display: 'flex',
-    flexWrap: 'wrap',
-    justifyContent: 'space-around',
     marginBottom: 20
-  },
-  gridList: {
-    width: '100%'
-  },
-  gridTile: {
-    cursor: 'pointer'
   },
   createButtonLarge: {
     position: 'fixed',
@@ -64,19 +52,10 @@ const styles = {
     width: '100%',
     display: 'inline-flex',
     marginBottom: 15
-  },
-  mapTypeIcon: {
-    marginLeft: 16,
-    marginRight: 16,
-    color: '#fff',
-    fontSize: '1.2rem'
-  },
-  mapTypeContainer: {
-    display: 'grid'
   }
 };
 
-export default class Maps extends Component {
+export default class Maps extends React.Component {
   constructor(props) {
     super(props);
     this.handleTabChange = this.handleTabChange.bind(this);
@@ -180,46 +159,10 @@ export default class Maps extends Component {
   renderMapContainer(maps) {
     if (maps.length > 0) {
       return (
-        <GridList
-          cols={this.props.large ? 4 : 2}
-          style={styles.gridList}
-          spacing={this.props.large ? 20 : 10}
-        >
-          {this.renderMaps(maps)}
-        </GridList>
+        <MapCollectionContainer maps={maps} />
       );
     } else {
       return this.renderNoMaps();
     }
-  }
-
-  renderMaps(maps) {
-    return maps.map(map => (
-      <GridListTile
-        key={map.id}
-        onClick={() => this.props.handleClickMap(map)}
-        style={styles.gridTile}
-      >
-        <img src={map.image_url} />
-        <GridListTileBar
-          title={map.name}
-          subtitle={<span>by: {map.owner_name}</span>}
-          actionIcon={<div style={styles.mapTypeContainer}>{this.renderMapTypeIcon(map)}</div>}
-        />
-      </GridListTile>
-    ));
-  }
-
-  renderMapTypeIcon(map) {
-    let actions = [];
-    if (map.private) {
-      actions.push(<LockIcon style={styles.mapTypeIcon} key="private" />);
-    }
-    if (map.shared) {
-      actions.push(<GroupIcon style={styles.mapTypeIcon} key="shared" />);
-    } else {
-      actions.push(<PersonIcon style={styles.mapTypeIcon} key="personal" />);
-    }
-    return actions;
   }
 }
